New Tree Data Structure
Using the OakTree class a Tree data structure was made from scratch with SetLeft(), SetRight(), GetLeft(), and Getright() methods. The set methods set an OakTree object (essentially a node) that contains a squirrel object as a branch of another OakTree object on either the right or left. The get methods return either the OakTree object to the right of the initial OakTree object or to the left of the initial OakTree object.